new_addressbook/backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.18.0
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
install==1.3.5
Jinja2==3.1.3
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.6.1
pydantic_core==2.16.2
python-jose==3.3.0
python-multipart==0.0.7
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
